webkit-efl/LayoutTests/platform/efl-tizen/fast/dom/Range/getClientRects-expected.txt

Test 1
FAIL rects.length should be 5. Was 6.
PASS rects[0].left is 8
PASS rects[0].top is 8
PASS rects[0].width is 400
FAIL rects[0].height should be 160. Was 200.
PASS rects[1].left is 8
FAIL rects[1].top should be 19. Was 18.
FAIL rects[1].width should be 396. Was 326.
FAIL rects[1].height should be 18. Was 19.
PASS rects[2].left is 8
FAIL rects[2].top should be 59. Was 58.
FAIL rects[2].width should be 398. Was 394.
FAIL rects[2].height should be 18. Was 19.
PASS rects[3].left is 8
FAIL rects[3].top should be 99. Was 98.
FAIL rects[3].width should be 360. Was 361.
FAIL rects[3].height should be 18. Was 19.
PASS rects[4].left is 8
FAIL rects[4].top should be 139. Was 138.
FAIL rects[4].width should be 306. Was 390.
FAIL rects[4].height should be 18. Was 19.
Test 2
FAIL rects.length should be 4. Was 5.
PASS rects[0].left is 8
FAIL rects[0].top should be 215. Was 256.
FAIL rects[0].width should be 396. Was 326.
FAIL rects[0].height should be 18. Was 19.
PASS rects[1].left is 8
FAIL rects[1].top should be 255. Was 296.
FAIL rects[1].width should be 398. Was 394.
FAIL rects[1].height should be 18. Was 19.
PASS rects[2].left is 8
FAIL rects[2].top should be 295. Was 336.
FAIL rects[2].width should be 360. Was 361.
FAIL rects[2].height should be 18. Was 19.
PASS rects[3].left is 8
FAIL rects[3].top should be 335. Was 376.
FAIL rects[3].width should be 306. Was 390.
FAIL rects[3].height should be 18. Was 19.
Test 3
FAIL rects.length should be 4. Was 5.
FAIL rects[0].left should be 80. Was 97.
FAIL rects[0].top should be 411. Was 494.
FAIL rects[0].width should be 324. Was 237.
FAIL rects[0].height should be 18. Was 19.
PASS rects[1].left is 8
FAIL rects[1].top should be 451. Was 534.
FAIL rects[1].width should be 398. Was 394.
FAIL rects[1].height should be 18. Was 19.
PASS rects[2].left is 8
FAIL rects[2].top should be 491. Was 574.
FAIL rects[2].width should be 360. Was 361.
FAIL rects[2].height should be 18. Was 19.
PASS rects[3].left is 8
FAIL rects[3].top should be 531. Was 614.
FAIL rects[3].width should be 88. Was 390.
FAIL rects[3].height should be 18. Was 19.
Test 4
FAIL rects.length should be 7. Was 8.
PASS rects[0].left is 8
FAIL rects[0].top should be 596. Was 722.
PASS rects[0].width is 400
FAIL rects[0].height should be 278. Was 318.
PASS rects[1].left is 8
FAIL rects[1].top should be 607. Was 732.
FAIL rects[1].width should be 396. Was 326.
FAIL rects[1].height should be 18. Was 19.
PASS rects[2].left is 8
FAIL rects[2].top should be 725. Was 772.
FAIL rects[2].width should be 242. Was 394.
FAIL rects[2].height should be 18. Was 19.
FAIL rects[3].left should be 326. Was 8.
FAIL rects[3].top should be 725. Was 890.
FAIL rects[3].width should be 44. Was 93.
FAIL rects[3].height should be 18. Was 19.
FAIL rects[4].left should be 8. Was 177.
FAIL rects[4].top should be 765. Was 890.
FAIL rects[4].width should be 341. Was 198.
FAIL rects[4].height should be 18. Was 19.
PASS rects[5].left is 8
FAIL rects[5].top should be 805. Was 930.
FAIL rects[5].width should be 366. Was 392.
FAIL rects[5].height should be 18. Was 19.
PASS rects[6].left is 8
FAIL rects[6].top should be 845. Was 970.
FAIL rects[6].width should be 67. Was 365.
FAIL rects[6].height should be 18. Was 19.
Test 5
FAIL rects.length should be 5. Was 6.
FAIL rects[0].left should be 80. Was 97.
FAIL rects[0].top should be 921. Was 1088.
FAIL rects[0].width should be 324. Was 237.
FAIL rects[0].height should be 18. Was 19.
PASS rects[1].left is 8
FAIL rects[1].top should be 1039. Was 1128.
FAIL rects[1].width should be 242. Was 394.
FAIL rects[1].height should be 18. Was 19.
FAIL rects[2].left should be 250. Was 8.
FAIL rects[2].top should be 950. Was 1246.
FAIL rects[2].width should be 76. Was 93.
FAIL rects[2].height should be 103. Was 19.
FAIL rects[3].left should be 326. Was 101.
FAIL rects[3].top should be 1039. Was 1158.
FAIL rects[3].width should be 44. Was 76.
FAIL rects[3].height should be 18. Was 103.
FAIL rects[4].left should be 8. Was 177.
FAIL rects[4].top should be 1079. Was 1246.
FAIL rects[4].width should be 284. Was 198.
FAIL rects[4].height should be 18. Was 19.
Test 6
FAIL rects.length should be 7. Was 9.
PASS rects[0].left is 8
FAIL rects[0].top should be 1224. Was 1434.
PASS rects[0].width is 400
FAIL rects[0].height should be 200. Was 280.
PASS rects[1].left is 8
FAIL rects[1].top should be 1235. Was 1444.
FAIL rects[1].width should be 396. Was 326.
FAIL rects[1].height should be 18. Was 19.
PASS rects[2].left is 8
FAIL rects[2].top should be 1275. Was 1484.
FAIL rects[2].width should be 57. Was 175.
FAIL rects[2].height should be 18. Was 19.
FAIL rects[3].left should be 105. Was 8.
FAIL rects[3].top should be 1275. Was 1524.
FAIL rects[3].width should be 44. Was 73.
FAIL rects[3].height should be 18. Was 19.
FAIL rects[4].left should be 8. Was 121.
FAIL rects[4].top should be 1315. Was 1524.
FAIL rects[4].width should be 387. Was 59.
FAIL rects[4].height should be 18. Was 19.
PASS rects[5].left is 8
FAIL rects[5].top should be 1355. Was 1564.
FAIL rects[5].width should be 397. Was 368.
FAIL rects[5].height should be 18. Was 19.
PASS rects[6].left is 8
FAIL rects[6].top should be 1395. Was 1604.
FAIL rects[6].width should be 171. Was 392.
FAIL rects[6].height should be 18. Was 19.
Test 7
FAIL rects.length should be 7. Was 9.
PASS rects[0].left is 8
FAIL rects[0].top should be 1471. Was 1762.
FAIL rects[0].width should be 396. Was 326.
FAIL rects[0].height should be 18. Was 19.
PASS rects[1].left is 8
FAIL rects[1].top should be 1500. Was 1802.
FAIL rects[1].width should be 400. Was 175.
FAIL rects[1].height should be 40. Was 19.
PASS rects[2].left is 8
FAIL rects[2].top should be 1511. Was 1832.
FAIL rects[2].width should be 57. Was 400.
FAIL rects[2].height should be 18. Was 40.
FAIL rects[3].left should be 105. Was 8.
FAIL rects[3].top should be 1511. Was 1842.
FAIL rects[3].width should be 44. Was 73.
FAIL rects[3].height should be 18. Was 19.
FAIL rects[4].left should be 8. Was 121.
FAIL rects[4].top should be 1551. Was 1842.
FAIL rects[4].width should be 387. Was 59.
FAIL rects[4].height should be 18. Was 19.
PASS rects[5].left is 8
FAIL rects[5].top should be 1591. Was 1882.
FAIL rects[5].width should be 397. Was 368.
FAIL rects[5].height should be 18. Was 19.
PASS rects[6].left is 8
FAIL rects[6].top should be 1631. Was 1922.
FAIL rects[6].width should be 171. Was 392.
FAIL rects[6].height should be 18. Was 19.
Test 8
FAIL rects.length should be 4. Was 5.
FAIL rects[0].left should be 168. Was 183.
FAIL rects[0].top should be 1678. Was 2058.
FAIL rects[0].width should be 269. Was 225.
FAIL rects[0].height should be 316. Was 263.
FAIL rects[1].left should be 137. Was 153.
FAIL rects[1].top should be 1704. Was 2084.
FAIL rects[1].width should be 271. Was 269.
FAIL rects[1].height should be 317. Was 315.
FAIL rects[2].left should be 107. Was 122.
FAIL rects[2].top should be 1729. Was 2110.
FAIL rects[2].width should be 246. Was 248.
FAIL rects[2].height should be 289. Was 290.
FAIL rects[3].left should be 76. Was 91.
FAIL rects[3].top should be 1755. Was 2136.
FAIL rects[3].width should be 212. Was 267.
FAIL rects[3].height should be 247. Was 312.
Test 9
FAIL rects.length should be 1. Was 0.
FAIL rects[0].left should be 8. Threw exception TypeError: 'undefined' is not an object (evaluating 'rects[0].left')
FAIL rects[0].top should be 1903. Threw exception TypeError: 'undefined' is not an object (evaluating 'rects[0].top')
FAIL rects[0].width should be 0. Threw exception TypeError: 'undefined' is not an object (evaluating 'rects[0].width')
FAIL rects[0].height should be 18. Threw exception TypeError: 'undefined' is not an object (evaluating 'rects[0].height')
Test 9b
PASS rects.length is 1
PASS rects[0].left is 8
FAIL rects[0].top should be 1903. Was 2318.
PASS rects[0].width is 0
FAIL rects[0].height should be 18. Was 19.
Test 10
FAIL rects.length should be 4. Was 5.
FAIL rects[0].left should be 19. Was 18.
FAIL rects[0].top should be 2088. Was 2546.
FAIL rects[0].width should be 18. Was 19.
FAIL rects[0].height should be 393. Was 380.
FAIL rects[1].left should be 59. Was 58.
FAIL rects[1].top should be 2088. Was 2546.
FAIL rects[1].width should be 18. Was 19.
FAIL rects[1].height should be 377. Was 394.
FAIL rects[2].left should be 99. Was 98.
FAIL rects[2].top should be 2088. Was 2546.
FAIL rects[2].width should be 18. Was 19.
FAIL rects[2].height should be 372. Was 361.
FAIL rects[3].left should be 139. Was 138.
FAIL rects[3].top should be 2088. Was 2546.
FAIL rects[3].width should be 18. Was 19.
FAIL rects[3].height should be 360. Was 390.
Test 11
PASS rects.length is 0
PASS successfullyParsed is true

TEST COMPLETE
